Use uint16 for storing count
A slice of uint16s fit more snugly in L1 cache and results in fewer misses.

Unfortunately, this limits the length of strings that can be used
to 65536.

Benchmark:
name              old time/op  new time/op  delta
Simple/ASCII-4     379ns ± 2%   330ns ± 2%  -13.03%  (p=0.000 n=10+10)
Simple/French-4    692ns ± 2%   617ns ± 2%  -10.75%  (p=0.000 n=10+10)
Simple/Nordic-4   1.28µs ± 1%  1.16µs ± 4%   -9.64%  (p=0.000 n=8+10)
Simple/Tibetan-4  1.16µs ± 3%  1.05µs ± 1%   -9.25%  (p=0.000 n=10+10)

+++ levenshtein.go
@@ -40,10 +40,10 @@ func ComputeDistance(a, b string) int {
 	lenS2 := len(s2)
 
 	// init the row
-	x := make([]int, lenS1+1)
+	x := make([]uint16, lenS1+1)
 	// we start from 1 because index 0 is already 0.
 	for i := 1; i < len(x); i++ {
-		x[i] = i
+		x[i] = uint16(i)
 	}
 
 	// make a dummy bounds check to prevent the 2 bounds check down below.
@@ -51,8 +51,8 @@ func ComputeDistance(a, b string) int {
 	_ = x[lenS1]
 	// fill in the rest
 	for i := 1; i <= lenS2; i++ {
-		prev := i
-		var current int
+		prev := uint16(i)
+		var current uint16
 		for j := 1; j <= lenS1; j++ {
 			if s2[i-1] == s1[j-1] {
 				current = x[j-1] // match
@@ -64,10 +64,10 @@ func ComputeDistance(a, b string) int {
 		}
 		x[lenS1] = prev
 	}
-	return x[lenS1]
+	return int(x[lenS1])
 }
 
-func min(a, b int) int {
+func min(a, b uint16) uint16 {
 	if a < b {
 		return a
 	}
